EN-79175: update pipelines library (#101)
This commit updates the version of the pipelines library used for the
Jenkinsfile. It also makes the necessary updates to the commonPipeline
parameters. Significant updates with the pipeline version upgrade
include using dockerWrapper rather than dockerize to manage the docker
build and also including security scanning.

The commit also updates the webhook for the team.